% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{build_graph}
\alias{build_graph}
\title{Build Collaborator Graph}
\usage{
build_graph(
  log,
  fileImportance,
  AgVP,
  AgCP,
  AbVP,
  AbCP,
  useFileImportance = TRUE,
  BFSW = 0.4,
  PCW = 0.6,
  UFSW = 0.15,
  BCW = 0.25,
  ECW = 0.3,
  DCW = 0.3,
  PCtimeBetweenPairProg = 30,
  PCtime_units = "days",
  FScollabTimespan = 7,
  FStime_units = "days",
  SCP = 0.5
)
}
\arguments{
\item{log}{Data.frame that contains: fileId, event, filePath, modifierStatus, programmerName, commitID, timestamp, programmerID}

\item{fileImportance}{dataframe with 2 columns "fileID" and "importance", optional argument}

\item{AgVP}{aggregation victim parameter: nodes with a weight < AgVP will be candidates for aggregation, the higher the AgVP the more clusters there will be. Value must be [0,1]}

\item{AgCP}{aggregation correlation parameter: a node will be clustered if it is a candidate (see AgVP) and has an edge with a distance significance > AgCP, the lower AgCP the higher the chance of clustering. Value must be [0,1]}

\item{AbVP}{abstraction victim parameter: nodes with weight < AbVP will be candidates for removal, the higher AbVP the more nodes can be removed. Value must be [0,1]}

\item{AbCP}{abstraction correlation parameter: nodes that are candidate for abstraction (see AbVP) will be abstracted if all their edges have a distance significance < AbCP, the higher the AbCP the more nodes will be removed. Value must be [0,1]}

\item{useFileImportance}{whether or not we want to use different importance values for the files. If set to FALSE but fileImportance arg was given, the fileImportance arg will be used nevertheless}

\item{BFSW}{binary frequency significance weight -> how much does the FS attribute to the edge weight}

\item{PCW}{proximity correlation weight -> how much does the PC attribute to the edge weight}

\item{UFSW}{unary frequency significance weight -> how much does the UFS attribute to the node weight}

\item{BCW}{betweenness centrality weight -> how much does the BC attribute to the node weight}

\item{ECW}{eigenvector centrality weight -> how much does the EC attribute to the node weight}

\item{DCW}{degree centrality weight -> how much does the DC attribute to the node weight}

\item{PCtimeBetweenPairProg}{for proximity correlation: param time that must be within 2 pair programming commits to count as a new pair programming session, default = 30}

\item{PCtime_units}{for proximity correlation: time unit of param timeBetweenPairProg, default "days"}

\item{FScollabTimespan}{for frequency significance : param timespan that indicates that 2 commits within this span are part of the same collaboration session (recursive), default = 7}

\item{FStime_units}{for frequency significance: time unit of param FScollabTimespan, default "days"}

\item{SCP}{SCP}
}
\value{
A graph object
}
\description{
Build Collaborator Graph
}
