% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collaborateR.R
\docType{package}
\name{collaborateR}
\alias{collaborateR}
\title{collaborateR - Create Collaboration Graphs from Version Control System Logs}
\description{
Create Collaboration Graphs from Version Control System Logs
}
