% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{create_igraph_object}
\alias{create_igraph_object}
\title{Function that creates an igraph object from the graph parameter}
\usage{
create_igraph_object(
  graph,
  anonymize = FALSE,
  nodeWeightScaling = 35,
  edgeWeightScaling = 50
)
}
\arguments{
\item{graph}{A graph, created with build_graph}

\item{anonymize}{Anonimize output}

\item{nodeWeightScaling}{Scale node weights}

\item{edgeWeightScaling}{Scale edge weights}
}
\value{
igraph object of the graph
}
\description{
Function that creates an igraph object from the graph parameter
}
