% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{read_vcs_eventlog}
\alias{read_vcs_eventlog}
\title{Read Version Control System Event log}
\usage{
read_vcs_eventlog(filePath)
}
\arguments{
\item{filePath}{Path to XES-file}
}
\value{
Data.frame that contains: fileId, event, filePath, modifierStatus, programmerName, commitID, timestamp, programmerID
}
\description{
XES-files with at least the following attributes

- Activity Type = "Commit"
- Time stamp = time stamp of the commit
- Resource = programmer that committed the file
}
