% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{vcs_log}
\alias{vcs_log}
\title{Version Control System event log}
\format{
## `vcs_log`
Event log consisting of 2423 rows and 7 columns:
\describe{
  \item{commitID}{a commit id}
  \item{fileId}{a file id}
  \item{programmerID}{id of a programmer}
  ...
}
}
\usage{
vcs_log
}
\description{
An example of a VCS event log
}
\keyword{datasets}
