% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeGraph.R
\name{visualize_graph}
\alias{visualize_graph}
\title{Function to visualize the graph}
\usage{
visualize_graph(
  graph,
  anonymize = FALSE,
  nodeWeightScaling = 35,
  edgeWeightScaling = 50
)
}
\arguments{
\item{graph}{the graph object to visualize}

\item{anonymize}{whether or not to anonymize node names}

\item{nodeWeightScaling}{scale the node weight by multiplying by this param}

\item{edgeWeightScaling}{scale the edge weight by multiplying by this param}
}
\value{
Visualized graph
}
\description{
Uses igraph's plot to plot the graph
}
