% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{write_graph_to_csv}
\alias{write_graph_to_csv}
\title{Write graph to 2 separate CSV files: one for the nodes, one for the edges}
\usage{
write_graph_to_csv(
  graph,
  nodesFilename = "nodes.csv",
  edgesFilename = "edges.csv",
  anonymization = FALSE
)
}
\arguments{
\item{graph}{: graph of which we want to write the nodes and edges to CSV}

\item{nodesFilename}{: filename for the CSV that will contain the nodes}

\item{edgesFilename}{: filename for the CSV that will contain the edges}

\item{anonymization}{: default FALSE meaning programmer names will not be anonymized}
}
\value{
CSV files
}
\description{
Write graph to 2 separate CSV files: one for the nodes, one for the edges
}
